// packages/client/src/components/domain/shop/OrderPage.module.css.d.ts
declare const styles:
  & Readonly<{ "main": string }>
  & Readonly<{ "productContainer": string }>
  & Readonly<{ "sub": string }>
  & Readonly<{ "square": string }>
  & Readonly<{ "title": string }>
   
  & Readonly<{ "description": string }>
  & Readonly<{ "price": string }>
  & Readonly<{ "stock": string }>
;
export default styles;
//# sourceMappingURL=./OrderPage.module.css.d.ts.map